Remove grandchildren mapping from TabbedShowLayout
Also fix linting.
