+++ packages/ra-ui-materialui/src/detail/TabbedShowLayoutTabs.js
@@ -1,25 +1,34 @@
-import React, { Component, Children, cloneElement } from "react";
-import PropTypes from "prop-types";
-import Tabs from "@material-ui/core/Tabs";
-import Divider from "@material-ui/core/Divider";
-import { withRouter, Route } from "react-router-dom";
-import compose from "recompose/compose";
-import { translate } from "ra-core";
+import React, { Children, cloneElement } from 'react';
+import PropTypes from 'prop-types';
+import Tabs from '@material-ui/core/Tabs';
 
-import CardContentInner from "../layout/CardContentInner";
+const getTabFullPath = (tab, index, baseUrl) =>
+    `${baseUrl}${
+      tab.props.path ? `/${tab.props.path}` : index > 0 ? `/${index}` : ''
+    }`;
 
-const TabbedShowLayoutTabs = ({ value, children, ...rest }) => (
-    <Tabs
-        value={value}
-        indicatorColor="primary"
-    >
-        {children}
+const TabbedShowLayoutTabs = ({ value, children, match, ...rest }) => (
+    <Tabs value={value} indicatorColor='primary'>
+        {Children.map(children, (tab, index) => {
+            if (!tab) return null;
+
+            // Builds the full tab tab which is the concatenation of the last matched route in the
+            // TabbedShowLayout hierarchy (ex: '/posts/create', '/posts/12', , '/posts/12/show')
+            // and the tab path.
+            // This will be used as the Tab's value
+            const tabPath = getTabFullPath(tab, index, match.url);
+
+            return cloneElement(tab, {
+                context: 'header',
+                value: tabPath,
+            });
+        })}
     </Tabs>
 );
 
 TabbedShowLayoutTabs.propTypes = {
     children: PropTypes.node,
-    value: PropTypes.number
+    value: PropTypes.number,
 };
 
 export default TabbedShowLayoutTabs;
